Docs: Split async reaction examples
