+++ docs/src/content/examples/abort-signals.mdx
@@ -1,0 +1,11 @@
+---
+title: 'Abort Signals'
+shortTitle: 'Abort Signals'
+sortIndex: 1
+exampleType: 'log'
+category: 'Reactivity'
+subcategory: 'Reactions'
+tags: ['reactivity', 'reaction', 'async', 'abort']
+description: Cancels in-flight work when a run is superseded
+tip: 'Pass `comp.abortSignal` to fetch or timers. It aborts when the run is invalidated or stopped'
+---
