+++ docs/src/content/examples/async-reactions.mdx
@@ -7,5 +7,5 @@ category: 'Reactivity'
 subcategory: 'Introduction'
 tags: ['reactivity', 'reaction', 'async', 'abort']
 description: Aborts stale runs and re-runs with the latest values
-tip: 'Runs never overlap. Mid-flight changes abort the run and coalesce into one re-run after settle'
+tip: 'Runs never overlap. A change mid-flight coalesces into one re-run after the current run settles'
 ---
